pull

+++ src/com/alipay/euler/andfix/patch/PatchManager.java
@@ -18,6 +18,7 @@
 package com.alipay.euler.andfix.patch;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
@@ -151,8 +152,12 @@ public class PatchManager {
 	public void addPatch(String path) throws IOException {
 		File src = new File(path);
 		File dest = new File(mPatchDir, src.getName());
+		if(!src.exists()){
+			throw new FileNotFoundException(path);
+		}
 		if (dest.exists()) {
-			mAndFixManager.removeOptFile(dest);
+			Log.d(TAG, "patch [" + path + "] has be loaded.");
+			return;
 		}
 		FileUtil.copyFile(src, dest);// copy to patch's directory
 		Patch patch = addPatch(dest);
@@ -206,8 +211,8 @@ public class PatchManager {
 			patchNames = patch.getPatchNames();
 			for (String patchName : patchNames) {
 				classes = patch.getClasses(patchName);
-				mAndFixManager.fix(patch.getFile(),
-						mContext.getClassLoader(), classes);
+				mAndFixManager.fix(patch.getFile(), mContext.getClassLoader(),
+						classes);
 			}
 		}
 	}
